File rename
file rename and a proptype validation.
